# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AssetCustomerSummary": {
        "properties": {
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "customer_type": {
            "title": "Customer Type",
            "type": "string"
          },
          "estimated_loss_kwh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Loss Kwh"
          },
          "external_ref": {
            "title": "External Ref",
            "type": "string"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "risk_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Tier"
          }
        },
        "required": [
          "customer_id",
          "external_ref",
          "customer_type",
          "risk_score",
          "risk_tier",
          "estimated_loss_kwh"
        ],
        "title": "AssetCustomerSummary",
        "type": "object"
      },
...
      "CustomerOut": {
        "properties": {
          "building_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Type"
          },
          "customer_type": {
            "title": "Customer Type",
            "type": "string"
          },
          "external_ref": {
            "title": "External Ref",
            "type": "string"
          },
          "feeder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feeder Id"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operator_id": {
            "title": "Operator Id",
            "type": "string"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "risk_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Tier"
          },
          "tariff_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tariff Type"
          },
          "transformer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transformer Id"
          }
        },
        "required": [
          "id",
          "operator_id",
          "external_ref",
          "transformer_id",
          "feeder_id",
          "region_id",
          "customer_type",
          "tariff_type",
          "building_type",
          "geometry",
          "risk_score",
          "risk_tier"
        ],
        "title": "CustomerOut",
        "type": "object"
      },
      "CustomerReadings": {
        "properties": {
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/MeterReadingOut"
            },
            "title": "Readings",
            "type": "array"
          },
          "unit": {
            "default": "kWh",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "readings"
        ],
        "title": "CustomerReadings",
        "type": "object"
      },
      "CustomerRiskProfile": {
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/CustomerOut"
          },
          "loss_attribution_share": {
            "title": "Loss Attribution Share",
            "type": "number"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "peer_comparison": {
            "items": {
              "$ref": "#/components/schemas/PeerComparisonPoint"
            },
            "title": "Peer Comparison",
            "type": "array"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskScoreOut"
          }
        },
        "required": [
          "customer",
          "risk",
          "peer_comparison",
          "loss_attribution_share",
          "notes"
        ],
        "title": "CustomerRiskProfile",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeterReadingOut": {
        "properties": {
          "consumption_kwh": {
            "title": "Consumption Kwh",
            "type": "number"
          },
          "current": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "power_factor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Factor"
          },
          "reading_quality": {
            "title": "Reading Quality",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "voltage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voltage"
          }
        },
        "required": [
          "timestamp",
          "consumption_kwh",
          "voltage",
          "current",
          "power_factor",
          "reading_quality",
          "source"
        ],
        "title": "MeterReadingOut",
        "type": "object"
      },
...
      "Page_AssetCustomerSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssetCustomerSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "Page[AssetCustomerSummary]",
        "type": "object"
      },
      "Page_CustomerOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CustomerOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "Page[CustomerOut]",
        "type": "object"
      },
...
      "PeerComparisonPoint": {
        "properties": {
          "customer_kwh": {
            "title": "Customer Kwh",
            "type": "number"
          },
          "expected_kwh": {
            "title": "Expected Kwh",
            "type": "number"
          },
          "peer_median_kwh": {
            "title": "Peer Median Kwh",
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "customer_kwh",
          "peer_median_kwh",
          "expected_kwh"
        ],
        "title": "PeerComparisonPoint",
        "type": "object"
      },
      "RiskComponents": {
        "properties": {
          "anomaly_score": {
            "title": "Anomaly Score",
            "type": "number"
          },
          "grid_imbalance_score": {
            "title": "Grid Imbalance Score",
            "type": "number"
          },
          "peer_score": {
            "title": "Peer Score",
            "type": "number"
          },
          "spatial_score": {
            "title": "Spatial Score",
            "type": "number"
          },
          "supervised_probability": {
            "title": "Supervised Probability",
            "type": "number"
          }
        },
        "required": [
          "supervised_probability",
          "anomaly_score",
          "grid_imbalance_score",
          "peer_score",
          "spatial_score"
        ],
        "title": "RiskComponents",
        "type": "object"
      },
      "RiskExplanation": {
        "properties": {
          "contribution": {
            "title": "Contribution",
            "type": "number"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "feature",
          "label",
          "contribution",
          "direction"
        ],
        "title": "RiskExplanation",
        "type": "object"
      },
      "RiskScoreOut": {
        "properties": {
          "components": {
            "$ref": "#/components/schemas/RiskComponents"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "estimated_loss_kwh": {
            "title": "Estimated Loss Kwh",
            "type": "number"
          },
          "estimated_loss_value": {
            "title": "Estimated Loss Value",
            "type": "number"
          },
          "explanations": {
            "items": {
              "$ref": "#/components/schemas/RiskExplanation"
            },
            "title": "Explanations",
            "type": "array"
          },
          "feature_version": {
            "title": "Feature Version",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inspection_priority": {
            "title": "Inspection Priority",
            "type": "number"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "number"
          },
          "risk_tier": {
            "title": "Risk Tier",
            "type": "string"
          },
          "scored_at": {
            "format": "date-time",
            "title": "Scored At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "entity_type",
          "entity_id",
          "scored_at",
          "risk_score",
          "risk_tier",
          "components",
          "confidence",
          "estimated_loss_kwh",
          "estimated_loss_value",
          "currency",
          "inspection_priority",
          "explanations",
          "model_version",
          "feature_version"
        ],
        "title": "RiskScoreOut",
        "type": "object"
      },
...
      "RouteRequest": {
        "properties": {
          "customer_ids": {
            "items": {
              "type": "string"
            },
            "title": "Customer Ids",
            "type": "array"
          }
        },
        "required": [
          "customer_ids"
        ],
        "title": "RouteRequest",
        "type": "object"
      },
      "RouteResponse": {
        "properties": {
          "ordered_customer_ids": {
            "items": {
              "type": "string"
            },
            "title": "Ordered Customer Ids",
            "type": "array"
          },
          "route_geometry": {
            "additionalProperties": true,
            "title": "Route Geometry",
            "type": "object"
          },
          "total_distance_km": {
            "title": "Total Distance Km",
            "type": "number"
          }
        },
        "required": [
          "ordered_customer_ids",
          "route_geometry",
          "total_distance_km"
        ],
        "title": "RouteResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Authoritative operational API for non-technical loss detection.",
    "title": "GridTrace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assets/{asset_id}/customers": {
      "get": {
        "operationId": "asset_customers_api_v1_assets__asset_id__customers_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AssetCustomerSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Customers",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/v1/customers": {
      "get": {
        "operationId": "list_customers_api_v1_customers_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "min_risk",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Risk"
            }
          },
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CustomerOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Customers",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{customer_id}": {
      "get": {
        "operationId": "get_customer_api_v1_customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Customer",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{customer_id}/readings": {
      "get": {
        "operationId": "get_readings_api_v1_customers__customer_id__readings_get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerReadings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Readings",
        "tags": [
          "customers"
        ]
      }
    },
    "/api/v1/customers/{customer_id}/risk-profile": {
      "get": {
        "operationId": "get_risk_profile_api_v1_customers__customer_id__risk_profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerRiskProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Risk Profile",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/api/v1/inspections/route": {
      "post": {
        "operationId": "build_route_api_v1_inspections_route_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Route",
        "tags": [
          "inspections"
        ]
      }
    },
...
  }
}